Add tests for extension
